Add auto-enter to the Bank PIN plugin
This commit updates the Bank PIN plugin with a new option to auto-enter
your PIN. If enabled, this feature will type the keys for your bank PIN.
This means the Keyboard Bankpin in the Bank plugin must also be enabled.

It doesn't type the PIN incredibly fast, but it still might be a
convenient way to be able to ignore the issue of typing in the bank PIN
in other bot plugins.