Improve the reliability of Always Make All
The Always Make All plugin would sometimes click some random location on
the screen rather than the widget it was supposed to click. This seems
to be caused by the iUtils helper plugin that was being used to invoke
mouse actions.

As of this commit the dependency on iUtils is removed and the desired
menu actions are now invoked directly through the client.