Initial commit of the Karambwan Fisher plugin
This commit adds the initial Karambwan Fisher plugin. It currently
fishes the Karambwan and banks them in Zanaris. It requires that the
last fairy ring destination be the Karambwan fishing area as it does not
yet have the ability to configure the fairy ring. That will hopefully be
added in a future version of the plugin.